pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "streamcab"
version = "0.1.0"
requires-python = ">=3.11"

[project.optional-dependencies]
producer = [
    "kafka-python==2.0.2",
    "pandas==2.2.2",
    "pyarrow==16.1.0",
]
dashboard = [
    "streamlit==1.37.1",
    "pandas==2.2.2",
    "pyarrow==16.1.0",
    "plotly==5.23.0",
]
trainer = [
    "pandas==2.2.2",
    "pyarrow==16.1.0",
    "scikit-learn==1.5.1",
    "joblib==1.4.2",
    "xgboost==2.1.1",
]
predictor = [
    "pandas==2.2.2",
    "pyarrow==16.1.0",
    "joblib==1.4.2",
    "xgboost==2.1.1",
]
scripts = [
    "httpx>=0.27",
]
dev = [
    "ruff>=0.9.0",
    "pyright>=1.1.390",
    "pyspark>=4.1.1",
]

[tool.hatch.build.targets.wheel]
packages = ["apps"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "C4", "RUF"]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["apps"]

[tool.pyright]
pythonVersion = "3.11"
typeCheckingMode = "basic"
include = ["apps", "scripts"]
exclude = ["**/__pycache__", "data"]
